requirements.txt
matplotlib==3.9.1
pandas==2.2.2
Pillow==10.3.0
plotly==5.22.0
bcrypt==4.1.3
pytz==2024.1
PyYAML==6.0.1
streamlit==1.36.0
streamlit_authenticator==0.3.2
streamlit_option_menu==0.3.13
